requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==1.7.1
sqlparse==0.5.3
tzdata==2024.2
